Refactor: code review
