--- app/src/main/res/drawable/add.xml
@@ -1,5 +1,0 @@
-<vector android:height="24dp" android:tint="#5A5858"
-    android:viewportHeight="24" android:viewportWidth="24"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="@android:color/white" android:pathData="M19,13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z"/>
-</vector>
