--- app/src/main/res/drawable/arrow_back_40dp.xml
@@ -1,5 +1,0 @@
-<vector android:autoMirrored="true" android:height="40dp"
-    android:tint="#C2C2C2" android:viewportHeight="24"
-    android:viewportWidth="24" android:width="40dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="@android:color/white" android:pathData="M20,11H7.83l5.59,-5.59L12,4l-8,8 8,8 1.41,-1.41L7.83,13H20v-2z"/>
-</vector>
