--- app/src/main/res/drawable/blue_radius.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-
-    android:shape="rectangle" android:padding = "15dp"
-    >
-    <solid android:color="@color/white"></solid>
-    <corners
-        android:bottomLeftRadius="10dp"
-        android:bottomRightRadius="10dp"
-        android:topLeftRadius="10dp"
-        android:topRightRadius="10dp"></corners>
-    <stroke android:width="1dp"
-        android:color="@color/blue"></stroke>
-</shape>
