--- app/src/main/res/drawable/blue_rectangular.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-
-    android:shape="rectangle" android:padding = "0dp"
-
-    >
-    <solid android:color="@color/white"></solid>
-    <corners
-        android:bottomLeftRadius="5dp"
-        android:bottomRightRadius="5dp"
-        android:topLeftRadius="5dp"
-        android:topRightRadius="5dp"
-        ></corners>
-    <stroke android:width="0.5dp"
-        android:color="@color/blue"></stroke>
-</shape>
