--- app/src/main/res/drawable/habbit_making_checkbox.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:state_checked="true"
-        android:drawable="@drawable/checked_check">
-        <shape android:shape="rectangle">
-            <solid android:color="@color/blue" />
-        </shape>
-
-    </item>
-
-
-    <item
-        android:state_checked="false"
-        android:drawable="@drawable/unchecked_check"
-
-        />
-
-</selector>
