--- app/src/main/res/drawable/radius.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-
-    android:shape="rectangle" android:padding = "15dp"
-    >
-    <solid android:color="#ffffff"></solid>
-    <corners
-        android:bottomLeftRadius="15dp"
-        android:bottomRightRadius="15dp"
-        android:topLeftRadius="15dp"
-        android:topRightRadius="15dp"></corners>
-    <stroke android:width="2dp"
-        android:color="#E5E7EB"></stroke>
-</shape>
