--- app/src/main/res/drawable/unchecked_check.xml
@@ -1,5 +1,0 @@
-<vector android:height="24dp" android:tint="@color/grey"
-    android:viewportHeight="24" android:viewportWidth="24"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="@android:color/white" android:pathData="M9,16.17L4.83,12l-1.42,1.41L9,19 21,7l-1.41,-1.41z"/>
-</vector>
