+++ app/src/main/res/values/colors.xml
@@ -7,6 +7,4 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
-    <color name="grey">#F3F4F6</color>
-    <color name="blue">#0094FF</color>
 </resources>
